+++ Backend-Bot-master/app/models/chat_message.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, TIMESTAMP, func, Text, ForeignKey, Boolean
+from sqlalchemy import BigInteger, String, TIMESTAMP, func, Text, ForeignKey, Boolean
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db import Base
@@ -7,11 +7,11 @@ from app.db import Base
 class ChatMessage(Base):
     __tablename__ = 'chat_messages'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    ride_id: Mapped[int | None] = mapped_column(Integer, ForeignKey('rides.id'), nullable=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    ride_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey('rides.id'), nullable=True)
     text: Mapped[str | None] = mapped_column(Text, nullable=True)
-    sender_id: Mapped[int | None] = mapped_column(Integer, ForeignKey('users.id'), nullable=True)
-    receiver_id: Mapped[int | None] = mapped_column(Integer, ForeignKey('users.id'), nullable=True)
+    sender_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=True)
+    receiver_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=True)
     message_type: Mapped[str | None] = mapped_column(String(50), nullable=True)
     attachments = mapped_column(JSONB, nullable=True)
     is_moderated: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
